RAG/01_Basic_RAG/requirements.txt
python-dotenv
llama-index
llama-index-vector-stores-qdrant 
llama-index-readers-file 
llama-index-embeddings-fastembed 
llama-index-llms-openai
llama-index-llms-groq
qdrant_client
fastembed
chainlit